backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
pydantic
numpy
scikit-learn
joblib
psycopg2-binary
passlib[bcrypt]
fpdf
openai
python-multipart
python-dotenv
sentence-transformers
torch>=1.11.0
sentence-transformers
nvidia-cusolver-cu12==11.6.1.9
spacy